Generation of Non-Amortizing Loans, at Scale.
Creation of functions capable of (1) generating loan terms for a non-amortizing loan (2) generating lender-perspective cash flows & IRR given a dictionary of terms for a non-amortizing loan. Functions are on a fixed time-convention -- they assume loan parameters are based on 365 days with a 365/360 accrual method.

Does not provide loan schedule. In the long-run, given a fixed-set of parameters, including amortization, functions should return:
1. Loan Terms
2. Loan Schedule
3. Cash flows from the lender perspective
4. Metadata useful for the calculation of fund waterfalls. Ex: 'net-funding' delineated as a separate value for use in leverage & pref calcs